+++ test/physics/surface_fluxes.jl
@@ -1,12 +1,14 @@
 @testset "Prescribed surface sensible heat fluxes" begin
     for Model in (PrimitiveDryModel, PrimitiveWetModel)
+        tmp_output_path = mktempdir(pwd(), prefix = "tmp_testruns_")  # Cleaned up when the process exits
 
         # prescribe ocean
         spectral_grid = SpectralGrid(trunc=31)
+        output = NetCDFOutput(spectral_grid, path=tmp_output_path)
         ocean_heat_flux = PrescribedOceanHeatFlux(spectral_grid)
         land_heat_flux = SurfaceLandHeatFlux(spectral_grid)
         surface_heat_flux = SurfaceHeatFlux(ocean=ocean_heat_flux, land=land_heat_flux)
-        model = Model(spectral_grid; surface_heat_flux)
+        model = Model(spectral_grid; surface_heat_flux, output)
         add!(model, SpeedyWeather.SurfaceFluxesOutput()...)
 
         simulation = initialize!(model)
@@ -17,7 +19,7 @@
         ocean_heat_flux = SurfaceOceanHeatFlux(spectral_grid)
         land_heat_flux = PrescribedLandHeatFlux(spectral_grid)
         surface_heat_flux = SurfaceHeatFlux(ocean=ocean_heat_flux, land=land_heat_flux)
-        model = Model(spectral_grid; surface_heat_flux)
+        model = Model(spectral_grid; surface_heat_flux, output)
         add!(model, SpeedyWeather.SurfaceFluxesOutput()...)
 
         simulation = initialize!(model)
@@ -27,13 +29,15 @@
 end
 
 @testset "Prescribed surface evaporative fluxes" begin
+    tmp_output_path = mktempdir(pwd(), prefix = "tmp_testruns_")  # Cleaned up when the process exits
 
     # prescribe ocean
     spectral_grid = SpectralGrid(trunc=31)
+    output = NetCDFOutput(spectral_grid, path=tmp_output_path)
     evaporative_flux_ocean = PrescribedOceanEvaporation(spectral_grid)
     evaporative_flux_land = SurfaceLandEvaporation(spectral_grid)
     surface_evaporation = SurfaceEvaporation(ocean=evaporative_flux_ocean, land=evaporative_flux_land)
-    model = PrimitiveWetModel(spectral_grid; surface_evaporation)
+    model = PrimitiveWetModel(spectral_grid; surface_evaporation, output)
     
     simulation = initialize!(model)
     set!(simulation.prognostic_variables.ocean.evaporative_flux, (λ, ϕ) -> ϕ > 0 ? 5e-5 : 0, model.geometry)
@@ -43,7 +47,7 @@ end
     evaporative_flux_ocean = SurfaceOceanEvaporation(spectral_grid)
     evaporative_flux_land = PrescribedLandEvaporation(spectral_grid)
     surface_evaporation = SurfaceEvaporation(ocean=evaporative_flux_ocean, land=evaporative_flux_land)
-    model = PrimitiveWetModel(spectral_grid; surface_evaporation)
+    model = PrimitiveWetModel(spectral_grid; surface_evaporation, output)
 
     simulation = initialize!(model);
     set!(simulation.prognostic_variables.land.evaporative_flux, (λ, ϕ) -> ϕ > 0 ? 5e-5 : 0, model.geometry)