+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Introduce Field for data on grid [#732](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/732)
 - Add Millenium and Century periods and allow conversion of Month and Year to seconds [#746](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/746)
 - Return UnicodePlot after run! if UnicodePlots.jl is loaded [#748](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/748)
 - git-auto-commit-action v6 [#747](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/747)